requirements.txt
aiocron==1.8
aiosignal==1.2.0
async-timeout==4.0.1
asyncio==3.4.3
attrs==21.2.0
backports.zoneinfo==0.2.1
certifi==2021.10.8
charset-normalizer==2.0.9
croniter==1.1.0
dnspython==2.1.0
frozenlist==1.2.0
idna==3.3
multidict==5.2.0
pymongo==4.0.1
python-dateutil==2.8.2
pytz-deprecation-shim==0.1.0.post0
requests==2.26.0
six==1.16.0
typing-extensions==4.0.1
tzdata==2021.5
tzlocal==4.1
urllib3==1.26.7
yarl==1.7.2
